+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `DurationCounterMetric` logged rate is per second and not anymore per millisecond.
+
 ## [0.18.1] - 2024-04-08
 
 ### Fixed